DynamicMMap: ensure iSize and WorkSpace are assigned values in size_t range
Even if file sizes were larger, it wouldn't make sense to have these
values larger than the system can mmap(2).

Explicitly declare them size_t to avoid misunderstanding the intended
range of these values in future code modifications (like adding large
file support).

(Unreliable places where some values are out of the expected range
could be detected with -Wconversion -Wsign-conversion, but there are
too many of them to use these warnings flags.)

+++ apt-pkg/contrib/mmap.h
@@ -39,7 +39,7 @@ class MMap
    protected:
 
    unsigned long Flags;
-   unsigned long iSize;
+   size_t iSize;
    void *Base;
 
    bool Map(FileFd &Fd);
@@ -53,7 +53,7 @@ class MMap
    // Simple accessors
    inline operator void *() {return Base;}
    inline void *Data() {return Base;}
-   inline unsigned long Size() {return iSize;}
+   inline size_t Size() {return iSize;}
 
    // File manipulators
    bool Sync();
@@ -79,7 +79,7 @@ class DynamicMMap : public MMap
    protected:
 
    FileFd *Fd;
-   unsigned long WorkSpace;
+   size_t WorkSpace;
    Pool *Pools;
    unsigned int PoolCount;
 
@@ -92,8 +92,8 @@ class DynamicMMap : public MMap
    inline unsigned long WriteString(string const S) {return WriteString(S.c_str(),S.length());}
    void UsePools(Pool &P,unsigned int const Count) {Pools = &P; PoolCount = Count;}
 
-   DynamicMMap(FileFd &F,unsigned long Flags,unsigned long RequestedWorkSpace = 2*1024*1024);
-   DynamicMMap(unsigned long Flags,unsigned long WorkSpace = 2*1024*1024);
+   DynamicMMap(FileFd &F,unsigned long Flags,size_t RequestedWorkSpace = 2*1024*1024);
+   DynamicMMap(unsigned long Flags,size_t WorkSpace = 2*1024*1024);
    virtual ~DynamicMMap();
 };
